tests/wpt/web-platform-tests/element-timing/image-clipped-svg.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Element Timing: observe image inside SVG with small dimensions</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
let beforeRender;
async_test(function (t) {
  const observer = new PerformanceObserver(
    t.step_func_done(function(entryList) {
      assert_equals(entryList.getEntries().length, 1);
      const entry = entryList.getEntries()[0];
      const index = window.location.href.lastIndexOf('/');
      const pathname = window.location.href.substring(0, index) +
          '/resources/circle.svg';
      checkElement(entry, pathname, 'my_svg', beforeRender);
      // Image size is 200x200 but SVG size is 100x100 so it is clipped.
      checkRect(entry, [0, 100, 0, 100]);
    })
  );
  observer.observe({entryTypes: ['element']});
  beforeRender = performance.now();
}, "Able to observe svg image.");
</script>
<style>
body {
  margin: 0;
}
</style>
<svg width="100" height="100">
  <image href='resources/circle.svg' elementtiming='my_svg'/>
</svg>
